+++ test/unit/lib/declarationHandlerTests.js
@@ -919,6 +919,105 @@ describe('declarationHandler', () => {
                 });
         });
 
+        it('should apply RouteMap path prefix fixes', () => {
+            const newDeclaration = {
+                parsed: true,
+                Common: {
+                    RouteMap: {
+                        rm1: {
+                            name: 'rm1',
+                            entries: [
+                                {
+                                    name: 33,
+                                    action: 'deny',
+                                    match: {
+                                        asPath: 'asPath1',
+                                        ipv4: {
+                                            address: {
+                                                prefixList: 'prefixList1'
+                                            },
+                                            nextHop: {
+                                                prefixList: 'prefixList1NextHop'
+                                            }
+                                        }
+                                    }
+                                },
+                                {
+                                    name: 44,
+                                    action: 'permit',
+                                    match: {
+                                        asPath: 'asPath2',
+                                        ipv6: {
+                                            address: {
+                                                prefixList: 'prefixList2'
+                                            },
+                                            nextHop: {
+                                                prefixList: 'prefixList2NextHop'
+                                            }
+                                        }
+                                    }
+                                }
+                            ]
+                        }
+                    }
+                }
+            };
+
+            const state = {
+                originalConfig: {
+                    Common: {}
+                },
+                currentConfig: {
+                    parsed: true,
+                    Common: {}
+                }
+            };
+
+            const declarationHandler = new DeclarationHandler(bigIpMock);
+            return declarationHandler.process(newDeclaration, state)
+                .then(() => {
+                    const routeMap = declarationWithDefaults.Common.RouteMap.rm1;
+                    assert.deepStrictEqual(
+                        routeMap,
+                        {
+                            name: 'rm1',
+                            entries: [
+                                {
+                                    name: 33,
+                                    action: 'deny',
+                                    match: {
+                                        asPath: '/Common/asPath1',
+                                        ipv4: {
+                                            address: {
+                                                prefixList: '/Common/prefixList1'
+                                            },
+                                            nextHop: {
+                                                prefixList: '/Common/prefixList1NextHop'
+                                            }
+                                        }
+                                    }
+                                },
+                                {
+                                    name: 44,
+                                    action: 'permit',
+                                    match: {
+                                        asPath: '/Common/asPath2',
+                                        ipv6: {
+                                            address: {
+                                                prefixList: '/Common/prefixList2'
+                                            },
+                                            nextHop: {
+                                                prefixList: '/Common/prefixList2NextHop'
+                                            }
+                                        }
+                                    }
+                                }
+                            ]
+                        }
+                    );
+                });
+        });
+
         it('should apply GSLB server fix', () => {
             const newDeclaration = {
                 parsed: true,