+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes to this project are documented in this file. More detail and links can b
 
 ## 1.19.0
 ### Added
+- AUTOTOOL-1990: Route Map
 - AUTOTOOL-2175: Add support for gslb monitor http
 - AUTOTOOL-2176: GSLB Prober Pool
 ### Fixed